requirements.txt
undetected-chromedriver>=3.5.5
selenium>=4.18.0
anthropic>=0.40.0
python-dotenv>=1.0.0
